Address some review comments Use string builder Avoid mutating supplied parameters Remove unused collex ID from DTO Move DTO into UI module
