+++ src/main/java/uk/gov/ons/ssdc/supporttool/utility/CaseSearchResultsMapper.java
@@ -8,7 +8,7 @@ import java.util.UUID;
 import lombok.SneakyThrows;
 import org.springframework.jdbc.core.RowMapper;
 import org.springframework.stereotype.Component;
-import uk.gov.ons.ssdc.supporttool.model.dto.CaseSearchResult;
+import uk.gov.ons.ssdc.supporttool.model.dto.ui.CaseSearchResult;
 
 @Component
 public class CaseSearchResultsMapper implements RowMapper<CaseSearchResult> {
@@ -21,9 +21,6 @@ public class CaseSearchResultsMapper implements RowMapper<CaseSearchResult> {
     caseContainerDto.setId(resultSet.getObject("id", UUID.class));
     caseContainerDto.setCaseRef(resultSet.getString("case_ref"));
     caseContainerDto.setSample(objectMapper.readValue(resultSet.getString("sample"), Map.class));
-
-    caseContainerDto.setCollectionExerciseId(
-        resultSet.getObject("collection_exercise_id", UUID.class));
     caseContainerDto.setCollectionExerciseName(resultSet.getString("collex_name"));
     return caseContainerDto;
   }
