+++ ui/src/SurveySampleSearch.js
@@ -49,7 +49,7 @@ class SurveySampleSearch extends Component {
       return;
     }
     this.setState({ searchTermFailedValidation: false });
-    let searchUrl = `searchInSurvey/${this.props.surveyId}?searchTerm=${this.state.searchTerm}`;
+    let searchUrl = `surveyCases/${this.props.surveyId}?searchTerm=${this.state.searchTerm}`;
 
     if (this.state.selectedCollectionExercise) {
       searchUrl += `&collexId=${this.state.selectedCollectionExercise}`;
